chore: update to revision d20250509
Acked-for-series: Varunkumar Allagadapa <varunkumar.allagadapa@amd.com>

+++ include/lib_common/IDriver.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: © 2024 Allegro DVT <github-ip@allegrodvt.com>
+// SPDX-FileCopyrightText: © 2025 Allegro DVT <github-ip@allegrodvt.com>
 // SPDX-License-Identifier: MIT
 
 /******************************************************************************
@@ -7,6 +7,8 @@
    \file
  *****************************************************************************/
 #pragma once
+
+#include "lib_rtos/types.h"
 
 typedef enum AL_EDriverError
 {
@@ -28,9 +30,9 @@ typedef enum AL_EDriverError
 typedef struct AL_TDriver AL_TDriver;
 typedef struct AL_TDriverVTable
 {
-  int (* pfnOpen)(AL_TDriver* driver, const char* device);
-  void (* pfnClose)(AL_TDriver* driver, int fd);
-  AL_EDriverError (* pfnPostMessage)(AL_TDriver* driver, int fd, long unsigned int messageId, void* data, bool isBlocking);
+  int32_t (* pfnOpen)(AL_TDriver* driver, const char* device);
+  void (* pfnClose)(AL_TDriver* driver, int32_t fd);
+  AL_EDriverError (* pfnPostMessage)(AL_TDriver* driver, int32_t fd, uint32_t messageId, void* data, bool isBlocking);
 }AL_TDriverVTable;
 
 struct AL_TDriver
@@ -39,25 +41,25 @@ struct AL_TDriver
 };
 
 static inline
-int AL_Driver_Open(AL_TDriver* driver, const char* device)
+int32_t AL_Driver_Open(AL_TDriver* driver, const char* device)
 {
   return driver->vtable->pfnOpen(driver, device);
 }
 
 static inline
-void AL_Driver_Close(AL_TDriver* driver, int fd)
+void AL_Driver_Close(AL_TDriver* driver, int32_t fd)
 {
   driver->vtable->pfnClose(driver, fd);
 }
 
 static inline
-AL_EDriverError AL_Driver_PostBlockingMessage(AL_TDriver* driver, int fd, long unsigned int messageId, void* data)
+AL_EDriverError AL_Driver_PostBlockingMessage(AL_TDriver* driver, int32_t fd, uint32_t messageId, void* data)
 {
   return driver->vtable->pfnPostMessage(driver, fd, messageId, data, true);
 }
 
 static inline
-AL_EDriverError AL_Driver_PostNonBlockingMessage(AL_TDriver* driver, int fd, long unsigned int messageId, void* data)
+AL_EDriverError AL_Driver_PostNonBlockingMessage(AL_TDriver* driver, int32_t fd, uint32_t messageId, void* data)
 {
   return driver->vtable->pfnPostMessage(driver, fd, messageId, data, false);
 }
